backend/requirements.txt
# Python dependencies for CS Girlies Hackathon project

# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Environment variables
python-dotenv==1.0.0

# AI/ML Services (choose what you need)
openai==1.3.0              # For OpenAI GPT
# google-generativeai      # For Google Gemini
# anthropic                # For Claude

# Database (choose based on your needs)
# pymongo                  # For MongoDB
# psycopg2-binary         # For PostgreSQL
# firebase-admin          # For Firebase

# Data processing
requests==2.31.0
# pandas                   # For data manipulation
# numpy                    # For numerical operations

# Utilities
gunicorn==21.2.0          # For production deployment
# pytest                   # For testing
# black                    # For code formatting
# flake8                   # For linting
